fxas21002c split out main and header
