+++ src/lib/drivers/barometer/PX4Barometer.hpp
@@ -31,6 +31,8 @@
  *
  ****************************************************************************/
 
+#pragma once
+
 #include <drivers/drv_baro.h>
 #include <drivers/drv_hrt.h>
 #include <lib/cdev/CDev.hpp>
