+++ src/lib/drivers/gyroscope/PX4Gyroscope.hpp
@@ -31,6 +31,7 @@
  *
  ****************************************************************************/
 
+#pragma once
 
 #include <drivers/device/integrator.h>
 #include <drivers/drv_gyro.h>
