+++ src/lib/drivers/magnetometer/PX4Magnetometer.hpp
@@ -31,6 +31,8 @@
  *
  ****************************************************************************/
 
+#pragma once
+
 #include <drivers/drv_mag.h>
 #include <drivers/drv_hrt.h>
 #include <lib/cdev/CDev.hpp>
